+++ src/app.tsx
@@ -2,7 +2,7 @@ import { ErrorBoundary } from 'react-error-boundary';
 import PrintError from '@/components/ErrorBoundary';
 // import LoginPage from '@/pages/login';
 // import LandingPage from '@/pages/landing-page';
-// import HobbySelectPage from '@/pages/hobby-selecet';
+// import HobbySelectPage from '@/pages/hobby-select';
 
 function App() {
   return (
